COSMIT refactor splitter data declaration and copying
Note that if we are willing to suffer a change to the
Splitter.node_split interface, node_split could take a SplitInfo*
instead of 6 separate arguments, and we could save many lines from
node_split and TreeBuilders.

If it's of great interest, we could hurry the changed Splitter interface
into 0.15. Ping @ogrisel